json server 연결 및 get, render
