+++ public/index.html
@@ -15,6 +15,7 @@
 	<script defer src="./assets/js/jquery.mCustomScrollbar.concat.min.js"></script>
 	<script defer src="./assets/js/bundle.js"></script>
 	<script defer src="./assets/js/scroll.js"></script>
+	<script defer src="./assets/js/works.js"></script>
 </head>
 <body>
 	<div id="wrap">
