+++ server.js
@@ -12,4 +12,4 @@ app.get('/layerpopup', (req, res) => {
 });
 
 app.use('/', router);
-app.listen(3000, () => console.log('Hello localhost:3000'));
+app.listen(4000, () => console.log('Hello localhost:4000'));
