+++ server/db.json
@@ -1,0 +1,42 @@
+{
+  "works": [
+    {
+      "id": 1,
+      "title": "대주제",
+      "list": [
+        {
+          "id": 1,
+          "title": "소주제",
+          "date": "19/01/01 12:12:12",
+          "Description": "content",
+          "labels": {
+            "yellow": {
+              "completed": false
+            },
+            "orange": {
+              "completed": false
+            },
+            "red": {
+              "completed": false
+            },
+            "blue": {
+              "completed": false
+            }
+          },
+          "checklist": [
+            {
+              "id": 1,
+              "content": "check1",
+              "completed": false
+            },
+            {
+              "id": 2,
+              "content": "check2",
+              "completed": false
+            }
+          ]
+        }
+      ]
+    }
+  ]    
+}
