+++ src/assets/js/main.js
@@ -1,12 +1,10 @@
 import add from './add';
 import popup from './popup';
 import dragdrop from './dragdrop';
-import works from './works';
 
 popup();
 dragdrop();
 add();
-works();
 
 (function () {
   const $timer = document.querySelector('.digital-time');
